Sketch of changes
To make this work, we have to move the computations in the graph, and
that requires changing the model primitive, so these changes are not
functional. However, the algorithm has to look more like what it's
presented here.
Once the changes in the model have been made, we can return to these
changes.