Split build methods in model
Algorithms like MAML require access to the symbolic loss and gradient
from the graph built in the model class to insert additional primitives,
so the method to build the gradients had to be split at symbolic level.
Additionally, the a build method was added to build the entire model and
its optimization in a single call.