Add build method in MAML
A method to build the graph for the MAML algorithm was added, although
it's still not complete. Only the updates per task have been
implemented, and the second order derivative is not yet added.
Also, further modifications were made in the Model class, and methods to
create placeholders from the sampler class were added.