Add MetaSampler
For MAML, we need meta batches of data, that is, a batch of batches.
Each regular batch represents a task, and the task is to learn to
classify a subset of classes from the dataset.
Each training instance in the regular batch is composed of all the
classes to learn in task.
Across the meta batch, we learn to classify regardless of the classes in
each particular batch.
This commit adds a MetaSampler and the MnistMetaSampler to use with the
MAML algorithms.