Use tensors in dense layer instead of variables
The method _set_tensors was added in the Model class to use the tensors
fed in the build_forward_pass instead of the varaibles created in the
layer.
Also, the build_compute_gradients method was refactor to not use an
optimizer but the tf.gradients function.
However, there's still one bug in the code that is making the forward
pass for update 2 not use the fast weights of update 1 but update 0, and
therefore, the result gradient with respect the fast weights of update 2
turns out to be zero, causing an error while building the computation
graph.