Replace layer by operations in _set_tensors
Currently this approach is the one working to use tensors instead of
variables.
Also training accuracy measurements have been added in the MAML
algorithm, and the labels have been modified to always be from 0 to the
number of tasks, so the classifier doesn't cheat in learning the
features of the MNIST digits and only focuses in learning how to
clasify.