Train VAE with MAML
VAE is working with MAML now, however, some tunning is required to
achieve a better loss.
Also, a change in the MNIST meta sampler has to be done to remove the
variable train_vae inside the MAML algorithm, since it should remain
agnostic to the model.
Another thing to fix is to not user hard coded scopes in the Model class
to get the trainable variables.
Also, dumping images in TensorBoard has to be improved.